refactor: code stlye
