+++ packages/shared/src/hooks/index.ts
@@ -2,6 +2,7 @@ export * from './useMenu.js'
 export * from './useRemarkable.js'
 export * from './useOpenShareTxDialog.js'
 export * from './useSnackbarCallback.js'
+export * from './useIterator.js'
 export * from './useIsImageURL.js'
 export * from './useTokenSecurity.js'
 export * from './useWeb3ProfileHiddenSettings.js'
